2011-06-28  Roland Steiner  <rolandsteiner@chromium.org>
        Reviewed by Tony Chang.

        Fix filesystem_unittest.FileSystemTest test_read_and_write_file()
        https://bugs.webkit.org/show_bug.cgi?id=63514

        - change write_text_file to write_binary_file
        - check that text_path and binary_path are actually existant files
        - change cleanup block from 'except' to 'finally'

        * Scripts/webkitpy/common/system/filesystem_unittest.py:

Canonical link: https://commits.webkit.org/79230@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@89944 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Tools/Scripts/webkitpy/common/system/filesystem_unittest.py
@@ -156,13 +156,13 @@ class FileSystemTest(unittest.TestCase):
             contents = fs.read_binary_file(text_path)
             self.assertEqual(contents, hex_equivalent)
 
-            fs.write_text_file(binary_path, hex_equivalent)
+            fs.write_binary_file(binary_path, hex_equivalent)
             text_contents = fs.read_text_file(binary_path)
             self.assertEqual(text_contents, unicode_text_string)
-        except:
-            if text_path:
+        finally:
+            if text_path and fs.isfile(text_path):
                 os.remove(text_path)
-            if binary_path:
+            if binary_path and fs.isfile(binary_path):
                 os.remove(binary_path)
 
     def test_read_binary_file__missing(self):
